requirements.txt
certifi==2018.1.18
chardet==3.0.4
futures==3.2.0
idna==2.6
pkg-resources==0.0.0
python-swiftclient==3.5.0
requests==2.18.4
six==1.11.0
urllib3==1.22
